Damage is now dealt every 10 consecutive hits. Reworked a some of the other scripts to calculate damage rather than score, since we have now decided to use a ranking based system from the time remaining rather than a damage-based scoring system. Off by one problem with the creator mode resolved as well

+++ Assets/Scripts/Objects/NoteScript.cs
@@ -54,10 +54,10 @@ public class NoteScript : MonoBehaviour {
         }
     }
 
-    public float destroyWithFeedback(GameObject hitArea, bool correct)
+    public int destroyWithFeedback(GameObject hitArea, bool correct)
     {
 
-        float score = 0;
+        int score = 0;
 
         if (hitArea == null) {
             feedback.GetComponent<PlayerFeedback>().GiveFeedback(100, correct);
